Refactor uniswap (#17)
* extract constants

* extract initialize in subclasses

* more float into Decimal

* extract `_swap` into subclasses

* move more functions

* even more clean up

* refactor uniswap v3

* clean up

* clean up

* lint

+++ alphaswarm/services/exchanges/jupiter/jupiter.py
@@ -33,7 +33,7 @@ class JupiterClient(DEXClient):
         self,
         base_token: TokenInfo,
         quote_token: TokenInfo,
-        quote_amount: float,
+        quote_amount: Decimal,
         slippage_bps: int = 100,
     ) -> SwapResult:
         """Execute a token swap on Jupiter (Not Implemented)"""
